+++ src/filetypes/bkdjvu.h
@@ -90,8 +90,8 @@ class BKDJVU : public BKDocument {
 	virtual int setRotation(int, bool bForce=false);
 
 	virtual bool isBookmarkable();
-	virtual void getBookmarkPosition(map<string, int>&);
-	virtual int setBookmarkPosition(map<string, int>&);
+	virtual void getBookmarkPosition(map<string, float>&);
+	virtual int setBookmarkPosition(map<string, float>&);
 	virtual float getCurrentZoom();
 
 	static BKDJVU* create(string& file,string& longfilename);
