+++ src/filetypes/bkmudocument.h
@@ -112,8 +112,8 @@ public:
 	virtual int setRotation(int, bool bForce=false);
 
 	virtual bool isBookmarkable();
-	virtual void getBookmarkPosition(map<string, int>&);
-	virtual int setBookmarkPosition(map<string, int>&);
+	virtual void getBookmarkPosition(map<string, float>&);
+	virtual int setBookmarkPosition(map<string, float>&);
 };
 
 #endif
